Add functionality to delete_iam_user (for monthly leavers) - Individual delete_iam_user script - For stale iam users, the logic will be moved into delete_inactive_iam_user.py - delete_inactive_iam_user.py will need refactoring (todo)
